libpressio version 0.88.3
Minor changes

+ added smoke tests.  Smoke tests are intended for lightweight checking
  if everything is built and installed correctly.  These are intended
  for use in E4S and other CI systems.
+ ndzip now properly exports pressio:lossless

Bug fixes

+ bug fixes for tests when building core mode
+ when csv support is not built, don't test it
+ wrong type for pressio_thread_safety in io, metrics, and launch
+ various bug fixes for cusz
+ various fixes for smoke tests
+ distributed_manager now won't cause an `exit` when MPI_Init hasn't
  been called
+ libpressio now exports in the build directory
+ fixed a bug where deallocation would fail in python because swig
  unloaded a module
+ SZ3 renamed a version variable

+++ src/plugins/io/select.cc
@@ -20,7 +20,7 @@ struct select_io: public libpressio_io_plugin {
     pressio_options opts;
     opts.copy_from(impl->get_configuration());
     set(opts, "pressio:stability",  "unstable");
-    set(opts, "pressio:thread_safe",  static_cast<int32_t>(pressio_thread_safety_single));
+    set(opts, "pressio:thread_safe",  pressio_thread_safety_single);
     return opts;
   }
