+++ src/plugins/metrics/entropy.cc
@@ -56,7 +56,7 @@ public:
   struct pressio_options get_configuration() const override {
     pressio_options opts;
     set(opts, "pressio:stability", "stable");
-    set(opts, "pressio:thread_safe", static_cast<int32_t>(pressio_thread_safety_multiple));
+    set(opts, "pressio:thread_safe", pressio_thread_safety_multiple);
     return opts;
   }
